Give the run-named files and functions distinct names
cmd/preflight had run.go, cmd_run.go, runRun, runCheck, runExec, runFile and
runCmd, meaning six different things. The worst pair was run.go and
cmd_run.go: they shared a stem but were unrelated, and the more general one
had the shorter name, so anyone asking how `preflight run` works opened
run.go and found the check-execution helper every command uses.

runRun was the handler convention collapsing on itself — every other one is
run<X>Check, which for a subcommand named run gives runRunCheck.

  run.go      -> runcheck.go       matches the runCheck it holds
  runRun      -> runPreflightFile  says what it does
  main.go     -> root.go           rootCmd, Version, reportExecuteError
              -> hashbang.go       the argv[1]-is-a-script path
              -> execmode.go       the -- passthrough
  main_entry.go -> main.go         main() now lives in main.go

Test files follow the same split. No behavior change; the suite, vet on all
three platforms, and a manual pass over run, the .preflight escape, quiet
failures, usage errors, the exec guard and hashbang mode all agree.
